Improve test structure
