+++ zrml/asset-router/src/lib.rs
@@ -20,7 +20,9 @@
 
 extern crate alloc;
 
+#[cfg(test)]
 mod mock;
+#[cfg(test)]
 mod tests;
 pub use pallet::*;
