+++ zrml/asset-router/src/tests/mod.rs
@@ -1,0 +1,31 @@
+// Copyright 2023 Forecasting Technologies LTD.
+//
+// This file is part of Zeitgeist.
+//
+// Zeitgeist is free software: you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the
+// Free Software Foundation, either version 3 of the License, or (at
+// your option) any later version.
+//
+// Zeitgeist is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
+
+#![cfg(test)]
+
+use super::{mock::*, Error};
+use frame_support::{assert_err, assert_ok, traits::tokens::fungibles::Create};
+use orml_traits::{
+    BalanceStatus, MultiCurrency, MultiCurrencyExtended, MultiLockableCurrency,
+    MultiReservableCurrency, NamedMultiReservableCurrency,
+};
+use zeitgeist_primitives::types::Assets;
+
+mod multi_currency;
+mod multi_lockable_currency;
+mod multi_reservable_currency;
+mod named_multi_reservable_currency;
