+++ zrml/neo-swaps/src/lib.rs
@@ -23,7 +23,9 @@ extern crate alloc;
 mod benchmarking;
 mod consts;
 mod math;
+#[cfg(test)]
 mod mock;
+#[cfg(test)]
 mod tests;
 pub mod traits;
 pub mod types;
